Updates most of the DOS related stuff to the new Python API.
This updates most of the stuff related to DOS to the new Python API.
There is a TODO present in the dos.py file after which the code has yet
to be updated.

The code has not been fully test run since no DOS data is available for
testing at this time.