Remove partial CC0-licensing from LICENSE file
The licensing of *everything* done 2018 under CC0 was due to a misunderstanding. The only thing that needs to be CC0 licensed is Inviwo modules that we may want to propagate upstream to them. In those cases, that double-licensing is simply indicated in the files themselves. Someone should go through also all non-C++ source files and remove the superfluous CC0 parts from the headers.